git-svn-id: http://financereport.googlecode.com/svn/FinancialReport@1121 d2ea8a1b-09b1-a710-b10d-4e27fa839f9d

+++ trunk/Java/DataLoadSpringAmar/src/com/pikefin/businessobjects/FactData.java
@@ -4,12 +4,12 @@ package com.pikefin.businessobjects;
 import java.math.BigDecimal;
 import java.util.Calendar;
 import java.util.Date;
-
+import java.lang.Double;
 
 public class FactData implements java.io.Serializable {
 
 	 private Integer factDataId;
-     private BigDecimal value;
+     private Double value;
      private Integer scale;
      private Boolean manualCorrection;
      private Date dateCollected;
@@ -32,10 +32,10 @@ public class FactData implements java.io.Serializable {
 	public void setFactDataId(Integer factDataId) {
 		this.factDataId = factDataId;
 	}
-	public BigDecimal getValue() {
+	public Double getValue() {
 		return value;
 	}
-	public void setValue(BigDecimal value) {
+	public void setValue(Double value) {
 		this.value = value;
 	}
 	public Integer getScale() {
